"ground detection" Improved
A “little big” trouble on the improvised collision detector was the
gravity persistence, there was not a mechanism that determine when to
stop bouncing. We corrected this, by comparing the differences between
the two last bouncing velocities. As result, we are a step closer to
discover how to implement a more sophisticated collision detection
engine.